% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate an SQL script}
\usage{
generate(
  input,
  table_names,
  output_path,
  output_file,
  server_path,
  generate_files = TRUE
)
}
\arguments{
\item{input}{A list of data frames or a vector of \code{.csv} file names.}

\item{table_names}{A string vector that contains the names of the tables that
should be created in the SQL database. There should be one element in the
vector for each data frame or \code{.csv} file provided in the \code{input}
argument.}

\item{output_path}{The path for the local directory where the generated SQL
script will be saved. The path should end with a \code{/} but the function
will add one if it does not.}

\item{output_file}{The name of the generated SQL script without a file
extension. The SQL script will be saved as a \code{.txt} file. If you do
provide an extension, it will be replaced with \code{.txt}.}

\item{server_path}{The path for the folder on the server where you will store
the \code{.csv} files to be added to the database. The path should end with
a \code{/} but the function will add one if it does not. This path is
required to generate the \code{INFLIE} command in the SQL script that will
read the data into the database.}

\item{generate_files}{Defaults to \code{TRUE}. If \code{TRUE}, the function
will create a \code{.csv} file for each dataset in the \code{output_path}
directory, which you can then transfer to your server. You need to put
these files in the \code{server_path} directory that you specify in order
for the script to work properly. The names of the files will be generated
automatically and will correspond exactly to the names of the tables that
will be created in the database.}
}
\value{
Writes a \code{.txt} file with SQL commands and (optionally)
  \code{.csv} files in the \code{output_file} directory.
}
\description{
This function generates an SQL script that you can use to read datasets into
an SQL or MariaDB database. The use case for this function is that you have
one or more datasets, stored as \code{.csv} files in a local directory, or
existing as data frames in your current \code{R} workspace, that you want to
quickly read into an existing SQL or MariaDB database on a remote server.
This function will generate an SQL script that creates a table for each
dataset and quickly reads each dataset into the database.
}
\details{
The SQL script will automatically overwrite without warning any table in the
database with the same name as one of the tables being created. You will get
a warning about this every time you run the function.

You will need to store a \code{.csv} file for each dataset on your server in
the \code{server_path} directory before you run the SQL script or you will
get SQL errors about not being able to find the files. The function generates
these \code{.csv} files for you if you have not already prepared them and
will properly convert \code{NA} values in \code{R} to \code{\\N} values in
the SQL database.

If any of your datasets include a variable whose name is a reserved word in
SQL, the function will automatically escape the variable name in the
generated script.
}
